# Conflicts:
#	source/Esf.Website/gulpfile.js
#	source/Esf.Website/package.json

+++ source/Esf.DataAccess.Deployment/DeleteDb.ps1
@@ -1,2 +1,5 @@
-$dbDeploymentConfig = (Get-Content .\dbDeploymentConfig.json) -join "`n" | ConvertFrom-Json
-Remove-Item $dbDeploymentConfig.esFiddleDbPath -recurse
+$dbDeploymentConfig = (Get-Content .\dbDeployment.config.json) -join "`n" | ConvertFrom-Json
+if (Test-Path $dbDeploymentConfig.esFiddleDbPath)
+{
+	Remove-Item $dbDeploymentConfig.esFiddleDbPath -recurse
+}
